Adding modules and Fix remove functions
